+++ src/plonetheme/derico/setuphandlers.py
@@ -104,6 +104,7 @@ class HiddenProfiles:
             "plonetheme.derico.upgrades:1009",
             "plonetheme.derico.upgrades:1010",
             "plonetheme.derico.upgrades:1011",
+            "plonetheme.derico.upgrades:1012",
         ]